net/mlx5/hws: fix possible action setter segmenation fault
When the maximum action combination in RX is used we can get
a segfault due to an incorrecrt max array size define.
This bug can happen on RX/TX or FDB in the most complex
cases.
Current max was set to 7, but actual max is:
Max TX: 8, Max RX: 10, Max FDB: 9

Fixes: f8c8a6d ("net/mlx5/hws: add action object")
Signed-off-by: Alex Vesker <valex@nvidia.com>
Reviewed-by: Erez Shitrit <erezsh@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>