testpmd: cleanup cleanly from signal
Do clean shutdown of testpmd when a signal is received.
Instead of having testpmd commit suicide when a signal
is received, do a clean shutdown. This fixes problem
where a signal could be received in the middle of a PMD
and then the signal handler would call PMD's close routine
which could cause a deadlock.

Fixes: d9a191a00e81 ("app/testpmd: fix quitting in container")
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>