doc: fix miss blank line in testpmd flow syntax doc
In flow syntax documentation, there is example for create pattern
template.

Before the example, miss a blank line causing it to look regular bold
text.
In addition, inside the example, it uses tab instead of spaces which
expand the indentation in one line.

This patch adds the blank line and replaces tab with spaces.

Fixes: 04cc665fab38 ("app/testpmd: add flow template management")
Cc: akozyrev@nvidia.com
Cc: stable@dpdk.org

Signed-off-by: Michael Baum <michaelba@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>