config/arm: make SoC-generic more generic
The non-generic soc has 1280 max-lcores and 32 max-numas, but the
implementer-generic still has 256 max-lcores and 4 max-numas, which may
result in restrictions on the use of binaries compiled by soc-generic
in these systems.

This patch changes the soc-generic max-lcores to 1280 and max-numas to
32 to make it more generic.

Also, because single numa is a special case of multiple numas, mark
soc-generic's numa flag as true.

Signed-off-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>