net/failsafe: Fix crash due to in-valid sub-device port id
Crash occurring while the DPDK secondary processes trying to probe the
tap-device, where tap-device is a sub-device of Fail-safe device.
Some-times we get in-valid sub-devices(with the in-valid port-id's),
due to which the IPC communication does not get response and causes the
communication failures b/w primary/secondary process.
So, need to validate the sub-device(tap) while secondary process probe in
the Fail-safe PMD, to avoid such issues.

Bugzilla Id: 1116

Signed-off-by: Madhuker Mythri <madhuker.mythri@oracle.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>