net/mlx5: fix available tag registers calculation for HWS
Before this patch, if two ports in separate switch domains
were probed by an application, the shared array of available TAG
registers was calculated incorrectly.

When the intersection of supported REG_C registers and
available TAG registers was calculated,
capabilities were checked against an index of the TAG array,
not the register stored under that index.

This patch fixes this behavior by comparing capabilities mask
against registers stored in the TAG array.
Available TAG registers calculation is also refactored
to simplify the code.

Fixes: 8a89038f40ca ("net/mlx5: provide available tag registers")
Cc: bingz@nvidia.com

Signed-off-by: Dariusz Sosnowski <dsosnowski@nvidia.com>
Reviewed-by: Bing Zhao <bingz@nvidia.com>
Acked-by: Ori Kam <orika@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>