net/nfp: using the correct PCI name to look for the firmware
In the description of nfp document, DPDK PMD supports that the PF
looks for a firmware file by the PCI name.

After this commit, it will looks for the PCI name format such as
pci-0000:04:00.0.nffw rather than pci-04:00.0.nffw, which will
keep consistent with what is written in the nfp.rst.

Signed-off-by: Peng Zhang <peng.zhang@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Niklas Söderlund <niklas.soderlund@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>