vhost: decrease log level for unimplemented request
This patch changes VHOST_USER_SET_VRING_ERR "not
implemented" log level from INFO to DEBUG.

Indeed, it is not required to support this message type,
and since version 7.1, QEMU sends this message for every
queue. It could confuse the end-user, thinking there is an
issue where there are not.

Implementing it could be an option, but it does nothing
special on QEMU side but displaying an error message.

Fixes: fd29c33b651a ("vhost: handle unsupported message types in functions")
Cc: stable@dpdk.org

Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>