telemetry: add u32 value type
Currently, 32-bit integer value only is signed in telemetry. The u32 data
can not be assigned to signed 32-bit integer. However, assigning to u64 is
very wasteful, after all, the buffer capacity of each transfer is limited.
So it is necessary for 'u32' data to add usigned 32-bit integer type and
container type.

Cc: stable@dpdk.org

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Acked-by: Morten Brørup <mb@smartsharesystems.com>
Acked-by: Chengwen Feng <fengchengwen@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>