net/bnx2x: fix warnings about rte_memcopy lengths
Bugfix: The vlan in the bulletin does not contain a VLAN header, only the
VLAN ID, so only copy 2 byte, not 4. The target structure has padding
after the field, so copying 2 byte too many is effectively harmless.
There is no need to backport this patch.

Added type casts where copying arrays to the offset of a first field in a
structure holding multiple fields, to avoid compiler warnings.

Bugzilla ID: 1146

Signed-off-by: Morten Brørup <mb@smartsharesystems.com>

v3:
* First patch in series.

Signed-off-by: 0-day Robot <robot@bytheb.org>