net/mlx5: add stub functions to null drv ops
There are several functions implementation that queries the drv type
to understand which fops function to use.
In case the type is DV, the function gets the concrete DV function and
calls it.
In case it’s not, the function returns an error.

The current implementation is not flexible enough and will not support
future concrete functions in a good way.

This patch adds more stub functions that include error handling and
are assigned to the null drv ops struct.
This allows the functions to always call the virtual function without
basing the decision on a specific drv type.

Signed-off-by: Asaf Penso <asafp@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>