cryptodev: increase default max nb queue pairs
Increase default maximum number of queue pairs to 64. For sw crypto
PMDs, the number of queue pairs can be increased as required. The
default value (which is set to 8), prevents usage of sw cryptodev PMDs
with larger number of cores.

Application would be allowed to specify the number of queue pairs it
wishes to use. Only the upper bound is changed from 8 to 64.

Allow PMDs such as crypto_armv8 to run on larger number of cores.

Signed-off-by: Anoob Joseph <anoobj@marvell.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>