usertools: add an argment for file prefix
Currently, the file prefix for DPDK runtime directory is 'rte' which was
fixed in the telemetry client script. The user had to modify the prefix
each time to run this script if the file prefix of application isn't 'rte'.
So this patch addes an optional argment for the file prefix, like,
"./dpdk-telemetry-client.py -f rte_xxx -s file_path"

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>